# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Intelligence Platform API
  version: 0.1.0-phase0
  description: >
    Phase 0 design artifact. Defines the target external API surface for the
    Clinical Intelligence Platform: document ingestion, hybrid retrieval,
    conversational AI, and analytics. All endpoints require an authenticated,
    tenant-scoped bearer token issued by the Identity & Access Service
    (see docs/architecture/06-security-compliance.md §3). No endpoint accepts
    or returns data across tenant boundaries.


    **Tenant identification**: the server URL's `tenant_slug` and the bearer
    token's tenant claim MUST match. A mismatch is rejected with `403
    tenant_mismatch` before any handler logic runs — this is a validated
    invariant, not an assumed one (see docs/design/phase-0-architecture-review.md
    finding D19).


    **Versioning**: breaking changes are introduced as a new path-versioned
    major (`/v2`), never as an in-place breaking change to `/v1`. A version
    scheduled for removal is marked with a `Deprecation` response header and a
    `Sunset` header carrying the removal date at least 12 months before
    removal; deprecation status is also listed in the platform changelog, not
    only in headers (docs/design/phase-0-architecture-review.md finding D16).


    **Idempotency**: all POST endpoints that create a resource accept an
    `Idempotency-Key` request header. Replaying the same key with the same
...


...


...
    (finding D21).
  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /documents:
    post:
      tags: [Ingestion]
      summary: Register a document for ingestion
      description: >
        Registers a document (already uploaded to the tenant's object storage
        prefix, or provided inline for small payloads) and enqueues it for
        the ingestion pipeline (docs/architecture/02-rag-hybrid-retrieval.md §1).
        For bulk EHR exports, prefer `POST /documents/batch`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentIngestionRequest'
      responses:
        '202':
          description: Accepted for asynchronous ingestion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    get:
      tags: [Ingestion]
      summary: List documents visible to the caller's scope
      parameters:
        - $ref: '#/components/parameters/PatientIdFilter'
        - $ref: '#/components/parameters/DocumentTypeFilter'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /documents/{documentId}:
    get:
      tags: [Ingestion]
      summary: Get document metadata and ingestion status
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived, tenant- and scope-bound token issued by the Identity &
        Access Service after OIDC/SAML federation with the tenant's IdP.
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PatientIdFilter:
      name: patientId
      in: query
      schema: { type: string, format: uuid }
    DocumentTypeFilter:
      name: documentType
      in: query
      schema: { type: string }
    PageCursor:
      name: cursor
      in: query
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Replaying the same key + body returns the original response instead of creating a duplicate.
      schema: { type: string }
  headers:
    RetryAfter:
      description: Seconds to wait before retrying (429 responses)
      schema: { type: integer }
    RateLimitLimit:
      description: Request quota for the current window
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in the current window
      schema: { type: integer }
    RateLimitReset:
      description: Unix timestamp when the window resets
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Caller's scope does not permit this action or resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or not visible to caller's scope)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: RFC 7807 Problem Details (docs/design/phase-0-architecture-review.md finding D21)
      properties:
        type: { type: string, description: URI identifying the error category }
        title: { type: string, description: Short, human-readable summary }
        status: { type: integer, description: HTTP status code }
        detail: { type: string, description: Human-readable explanation specific to this occurrence }
        instance: { type: string, description: URI identifying this specific occurrence }
    DocumentIngestionRequest:
      type: object
      required: [objectStorageUri, documentType, sourceSystem]
      properties:
        objectStorageUri: { type: string }
        documentType:
          type: string
          enum: [discharge_summary, lab_report, radiology_note, trial_protocol,
                 adverse_event_report, literature, guideline, hl7v2_message,
                 fhir_bundle, dicom_study]
        sourceSystem: { type: string }
        patientId: { type: string, format: uuid }
        effectiveDate: { type: string, format: date }
    Document:
      type: object
      properties:
        documentId: { type: string, format: uuid }
        documentType: { type: string }
        ingestionStatus:
          type: string
          enum: [pending, parsed, extracted, embedded, graph_indexed, failed]
        deidentificationStatus:
          type: string
          enum: [not_deidentified, safe_harbor, expert_determination]
        createdAt: { type: string, format: date-time }
    DocumentList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Document' }
        nextCursor: { type: string, nullable: true }
...